requirements.txt
# Core
numpy
pandas

# Technical Analysis
ta

# Data
requests
ccxt

# AI / NLP (optional — system works without these)
transformers
sentence-transformers
torch
optuna
scipy

# Config
python-dotenv
pyyaml

# Forecasting and ML models (optional)
lightgbm
pytorch-forecasting
fingpt

# Testing
pytest

# Wavelet support for indicators
PyWavelets>=1.3.0
